# docs/openapi.yaml
openapi: 3.1.0
info:
  title: VigiaApp API
  version: 1.1.0
  description: |
    API REST de VigiaApp. Parte de la especificación v1.0 (`VigiaApp/docs/especificacion_api_db.md`);
    las diferencias con esa versión están en `docs/CAMBIOS_CLIENTE.md`.

    - Todas las respuestas son `application/json; charset=utf-8` (RNF-04).
    - Claves en `snake_case`; los booleanos viajan como `true`/`false` (RNF-07).
    - Fechas con hora: ISO-8601 en UTC con `Z` y sin milisegundos (`2026-09-23T18:35:00Z`).
      Fechas de calendario: `AAAA-MM-DD`. Horas: `HH:MM`.
    - Todas las rutas, salvo `/auth/login`, `/auth/clave` y `/auth/recuperar`, exigen `Authorization: Bearer <token>`.
    - Cualquier respuesta que no sea 2xx trae el cuerpo `Error`.
servers:
  - url: https://{dominio}/api/v1
    variables:
      dominio:
        default: api.vigiaapp.example
  - url: http://localhost:3000/api/v1
    description: Desarrollo local
security:
  - bearer: []
...
paths:
...
  /yo:
    get:
      tags: [Perfil]
      summary: Vecino de la sesión
      description: La app lo llama al abrir para validar el token guardado (401 → volver al login).
      responses:
        '200':
          description: Perfil propio
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Perfil' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
  /yo/clave:
    post:
      tags: [Perfil]
      summary: Cambiar la contraseña propia
      description: Cierra las demás sesiones del vecino; la actual sigue válida. Si `actual` no coincide responde 400 (no 401, para no cerrar la sesión en la app).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [actual, nueva]
              properties:
                actual: { type: string }
                nueva: { type: string, minLength: 8, maxLength: 128 }
      responses:
        '204': { description: Contraseña cambiada. }
        '400': { $ref: '#/components/responses/DatosInvalidos' }
        '401': { $ref: '#/components/responses/NoAutorizado' }

  # ---------------------------------------------------------------- Comité
  /manzanas:
    get:
      tags: [Comité]
      summary: Manzanas del barrio (cualquier vecino)
      responses:
        '200':
          description: Lista ordenada por nombre
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Manzana' } }
        '401': { $ref: '#/components/responses/NoAutorizado' }
    post:
      tags: [Comité]
      summary: Crear una manzana
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre]
              properties:
                nombre: { type: string, maxLength: 50 }
      responses:
        '201':
          description: Creada
          content: { application/json: { schema: { $ref: '#/components/schemas/Manzana' } } }
        '400': { $ref: '#/components/responses/DatosInvalidos' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403': { $ref: '#/components/responses/Prohibido' }
  /vecinos:
    get:
      tags: [Comité]
      summary: Todos los vecinos (activos e inactivos)
      responses:
        '200':
          description: Ordenados por apellido y nombre
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/VecinoAdmin' } }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403': { $ref: '#/components/responses/Prohibido' }
    post:
      tags: [Comité]
      summary: Dar de alta un vecino y enviarle la invitación por correo
      description: |
        El vecino queda con `invitacion_pendiente: true` hasta que usa el enlace. Si el correo falla, el
        vecino igual queda creado y la respuesta trae `invitacion_enviada: false` (se puede reenviar).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre, apellido, email, manzana_id]
              properties:
                nombre: { type: string, maxLength: 40 }
                apellido: { type: string, maxLength: 40 }
                email: { type: string, format: email, maxLength: 254 }
                telefono: { type: [string, 'null'], description: 'E.164, p. ej. +5491100000001' }
                manzana_id: { type: integer }
                rol: { $ref: '#/components/schemas/Rol' }
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/VecinoAdmin'
                  - type: object
                    required: [invitacion_enviada]
                    properties:
                      invitacion_enviada: { type: boolean }
        '400': { $ref: '#/components/responses/DatosInvalidos' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403': { $ref: '#/components/responses/Prohibido' }
        '409':
          description: Ya existe un vecino con ese email o teléfono.
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /vecinos/{id}:
    patch:
      tags: [Comité]
      summary: Editar o desactivar un vecino
      description: Solo se cambian los campos enviados. `activo:false` corta sus sesiones. El comité no puede cambiar su propio rol ni desactivarse (409).
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                nombre: { type: string, maxLength: 40 }
                apellido: { type: string, maxLength: 40 }
                telefono: { type: [string, 'null'] }
                manzana_id: { type: integer }
                rol: { $ref: '#/components/schemas/Rol' }
                activo: { type: boolean }
      responses:
        '200':
          description: Vecino actualizado
          content: { application/json: { schema: { $ref: '#/components/schemas/VecinoAdmin' } } }
        '400': { $ref: '#/components/responses/DatosInvalidos' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403': { $ref: '#/components/responses/Prohibido' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '409': { $ref: '#/components/responses/Conflicto' }
  /vecinos/{id}/invitacion:
    post:
      tags: [Comité]
      summary: Reenviar la invitación
      description: Invalida el enlace anterior. Si el vecino ya activó su cuenta responde 409 (que use "olvidé mi contraseña").
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '204': { description: Invitación enviada. }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403': { $ref: '#/components/responses/Prohibido' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '409': { $ref: '#/components/responses/Conflicto' }

  # ---------------------------------------------------------------- Inicio
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: 'Token de 64 hex que devuelven `/auth/login` y `/auth/clave`. Vale 24 h (RNF-03).'
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
  responses:
...
    DatosInvalidos:
      description: Falta un campo o tiene formato incorrecto.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { codigo: datos_invalidos, mensaje: 'presente debe ser true o false.' } }
    NoAutorizado:
      description: Token ausente, inválido o vencido (o credenciales incorrectas en /auth/login).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { codigo: no_autorizado, mensaje: 'Sesión vencida. Volvé a iniciar sesión.' } }
    Prohibido:
      description: El vecino no tiene permiso (p. ej. no es del comité).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { codigo: prohibido, mensaje: 'Solo el comité puede hacer esto.' } }
    NoEncontrado:
      description: El recurso no existe.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { codigo: no_encontrado, mensaje: 'La alerta no existe.' } }
    Conflicto:
      description: El estado actual no permite la operación.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { codigo: conflicto, mensaje: 'La alerta ya no está activa.' } }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [codigo, mensaje]
          properties:
            codigo:
              type: string
              enum: [datos_invalidos, no_autorizado, prohibido, no_encontrado, conflicto, demasiados_intentos, error_interno]
            mensaje: { type: string, description: 'Texto en castellano, apto para mostrar al usuario.' }
    Rol:
      type: string
      enum: [vecino, vigia, comite]
      default: vecino
...
    Perfil:
      type: object
      required: [id, nombre, apellido, email, telefono, manzana_id, manzana, rol]
      properties:
        id: { type: integer }
        nombre: { type: string }
        apellido: { type: string }
        email: { type: string, format: email }
        telefono: { type: [string, 'null'] }
        manzana_id: { type: integer }
        manzana: { type: string }
        rol: { $ref: '#/components/schemas/Rol' }
    VecinoAdmin:
      allOf:
        - $ref: '#/components/schemas/Perfil'
        - type: object
          required: [activo, invitacion_pendiente, ultimo_acceso]
          properties:
            activo: { type: boolean }
            invitacion_pendiente: { type: boolean, description: 'true = todavía no eligió contraseña' }
            ultimo_acceso: { type: [string, 'null'], format: date-time }
    Manzana:
      type: object
      required: [id, nombre]
      properties:
        id: { type: integer }
        nombre: { type: string, example: Manzana 4 }
...